Moving create_user into utility module; standardizing test names
